third_party/WebKit/LayoutTests/editing/deleting/delete-ligature-002.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_not_equals(window.eventSender, undefined,
        'This test requires window.eventSender');
    assert_selection(
        '<div contenteditable>\u0E27\u0E31|</div>',
        selection => {
            selection.document.defaultView.focus();
            eventSender.keyDown('Backspace', null);
        },
        '<div contenteditable>\u0E27|</div>');
    }, 'Backspace key should delete only last character of \u0E27\u0E31');
</script>
